Move localStorage related functions to db.js. Render cards dynamically with the data stored in localStorage
