+++ create.html
@@ -10,7 +10,13 @@
   <body>
     <form>
       <input type="text" name="title" id="title" placeholder="Enter title" />
-      <textarea name="tags" id="tags" cols="30" rows="10"></textarea>
+      <textarea
+        name="tags"
+        id="tags"
+        cols="30"
+        rows="10"
+        placeholder="Tags separated by comma"
+      ></textarea>
       <button type="submit">Create</button>
     </form>
   </body>
