+++ create.js
@@ -1,3 +1,5 @@
+import { addCard } from "./lib/db.js";
+
 const form = document.querySelector("form");
 
 form.addEventListener("submit", (event) => {
@@ -14,19 +16,7 @@ form.addEventListener("submit", (event) => {
 
   console.log(cardData);
 
-  let cards = JSON.parse(localStorage.getItem("cards"));
-  if (cards === null) {
-    cards = [];
-  }
-  cards.push(cardData);
-
-  try {
-    localStorage.setItem("cards", JSON.stringify(cards));
-    throw new Error("An error happened");
-  } catch (error) {
-    console.log(error);
-    alert("There was an error while saving");
-  }
+  addCard(cardData);
 
   form.reset();
 });
