+++ lib/db.js
@@ -1,32 +1,20 @@
-export const data = [
-  {
-    title: "Movies",
-    items: [
-      "The Lord of the Rings",
-      "The Matrix",
-      "Blade Runner",
-      "Tenet",
-      "Joker",
-      "Nomadland",
-    ],
-  },
-  {
-    title: "Series",
-    items: [
-      "Narcos",
-      "The Serpent",
-      "The Soprano",
-      "The Wire",
-      "True Detective",
-    ],
-  },
-  {
-    title: "Books",
-    items: [
-      "The Lord of the Rings",
-      "The Hobbit",
-      "Pride and Prejudice",
-      "Hamlet",
-    ],
-  },
-];
+export function getCards() {
+  let cards = JSON.parse(localStorage.getItem("cards"));
+  if (cards === null) {
+    cards = [];
+  }
+
+  return cards;
+}
+
+export function addCard(cardData) {
+  const cards = getCards();
+  cards.push(cardData);
+
+  try {
+    localStorage.setItem("cards", JSON.stringify(cards));
+  } catch (error) {
+    console.log(error);
+    alert("There was an error while saving");
+  }
+}
